requirements.txt
accelerate==1.6.0
av==12.1.0
decord
diffusers==0.33.1
einops
ffmpeg-python==0.2.0
gradio==5.25.2
imageio-ffmpeg==0.4.8
imageio==2.31.1
jinja2>=3.1.2
numpy==1.26.2
opencv-contrib-python
peft
pillow==11.1.0
requests==2.31.0
safetensors
scipy==1.12.0
sentencepiece==0.2.0
torchsde==0.2.6
tqdm
timm
transformers==4.46.2

# for toolbox
basicsr
# basicsr-fixed
devicetorch
facexlib>=0.2.5
gfpgan>=1.3.5
psutil
realesrgan
colorlog
